wbn-sign: Improve readability by splitting bloated files
