+++ js/sign/src/utils/constants.ts
