+++ js/sign/src/utils/read-types.d.ts
